Move RJSF Json logic helper
